Finding Depth of Binary Tree using recursion
Easiest way to determine how deep a binary tree is to use recursion. It's intuitive at this level since it's just a comparison of left vs right node depth. Runtime is O(n). Even though it's a binary tree, we have to traverse every node to get it's depth (we assume it's not balanced). Best case for space complexity is O(log n) if it's balanced since the recursion is evenly split. However, it can be O(n) if the tree looks like a linked list.

Problem statement can be found here: https://leetcode.com/problems/maximum-depth-of-binary-tree/