Solution to problem whether there's a path from root to a leaf for a target sum
This solution traverses all nodes in the matrix in worst case scenario to determine whether there exists an array/list of values that total a target number, hence, it's O(n) time complexity. It uses a call stack to calculate paths on the left and right children of a node, hence, space complexity is O(h) where h is the height of the three. It is not always guaranteed to be O(log n) especially if the tree is not balanced. In the worst case, space complexity is O(n) when the tree looks like a linked list. Complete problem statement can be found here: https://leetcode.com/problems/path-sum