Solution for obtaining a balanced tree representation of a sorted array
Since the list is sorted, we can divide the list into medians where we add the median as root, then add numbers less than the median as left child nodes, those greater than the median as right child nodes. Time complexity is O(n) as we traverse all elements. Space complexity is O(n) for the tree but we also use O(log n) for the call-stack and as we create left and then right sub nodes