Flattening a Binary Tree to a Linked List
To flatten a binary tree, we need to move to prune its child nodes and just keep one. In this case, we'll keep the right vertices and move the ones on the left to the right. We can do so using a series of recursive calls which take O(n) is tree is mostly balanced. Space complexity is O(n) since we have to use the call stack on every node in the linked list.