Finding common letters in a list of words
The answer shows the minimum number of occurence of each letter across all elements in the list. The time complexity is O(n*m) where n is the number of words in the list and m is the average length of each word. Space complexity is O(m) for the dictionary/hashmap.